+++ examples/SimpleChat/lib/controllers/main.ex
@@ -10,9 +10,9 @@ defmodule Simplechat.Main.Login do
     {:render, [], []}
   end
 
-  def join([{"username", username}]) do
-  	set_session_val(:username, username)
-    {:json, [result: :ok, username: username], [{"Content-Type", "application/json"}]}
+  def join([username: username]) do
+    set_session_val(:username, username)
+    {:json, [result: :ok, username: username], []}
   end
 
 end