+++ lib/weber/cowboy.ex
@@ -21,8 +21,12 @@ defmodule Cowboy do
     {_, port}      = :lists.keyfind(:http_port, 1, web_server_config)
     {_, acceptors} = :lists.keyfind(:acceptors, 1, web_server_config)
     {_, ssl}     = :lists.keyfind(:ssl, 1, web_server_config)
-        
-    dispatch = :cowboy_router.compile([{:_, [{:_, Handler.WeberReqHandler, name}]}])
+
+    {:ws, ws_config} = :lists.keyfind(:ws, 1, config)
+    {_, ws_mod}  = :lists.keyfind(:ws_mod, 1, ws_config)
+          
+    dispatch = :cowboy_router.compile([{:_, [{'/_ws', Handler.WeberWebSocketHandler, {name, ws_mod}}, 
+                                             {'/[...]', Handler.WeberReqHandler, name}]}])
 
     case ssl do
       true ->